Fix User, Teacher, Subject can be deleted and leave app in invalid state
- User cannot be deleted if a subject/grades registration exist
- Teacher cannot be deleted if it has a subject assignation
- Subject cannot be deleted if a teacher is still assigned or any student registration exists

+++ universitysystem/src/main/java/com/practice/universitysystem/service/subject/SubjectService.java
@@ -81,6 +81,12 @@ public class SubjectService {
     }
 
     public void deleteSubject(String name) {
+        if (!registrationRepository.findAllBySubjectId(getSubject(name).getId()).isEmpty()) {
+            throw new IllegalStateException("Cannot delete subject if it has any existing registration");
+        }
+        if (!getAllTeachers(name).isEmpty()) {
+            throw new IllegalStateException("Cannot delete subject if any teacher is still assigned");
+        }
         subjectRepository.delete(getSubject(name));
     }
